Have pcap_open_dead_with_tstamp_precision().
As with the open_offline routines, so with pcap_open_dead(); instead of
a nanosecond-resolution version, have a
pcap_open_dead_with_tstamp_precision() routine that takes a
PCAP_TSTAMP_PRECISION_ value, and make pcap_open_dead() a wrapper that
requests microsecond resolution.

Declare it in pcap/pcap.h while we're at it.