Remove Linux-specific time stamp precision stuff from pcap.c.
Do time stamp precisions the way we do time stamp types - if a module
supports more than one time stamp precision, have it attach a table of
time stamp precisions to the pcap_t, and have the code to set the time
stamp check that list.

Check for the two #ifdefs in pcap-linux.c to see whether it supports
nanosecond time stamps - and don't use the ioctl or the socket option
for those #ifdefs if they're not defined.

This means we don't need to check for it in the configure script.

Also, have both the time stamp type-setting and precision-setting
routines allow the default value (PCAP_TSTAMP_NONE/PCAP_TSTAMP_PRECISION_MICRO)
if no list is supplied.

Make "can't set the precision" an error; not being able to set the time
stamp *type* doesn't mean your code won't work at all, it just means the
time stamps won't be as accurate as you'd like, but not being able to
set the time stamp *precision* means you'll be getting
seconds/microseconds, and your code may have to cope with that.