Just return PCAP_ERROR_NOT_ACTIVATED for some calls.
Return PCAP_ERROR_NOT_ACTIVATED, not PCAP_ERROR (-1), for calls to
pcap_datalink(), pcap_datalink_ext(), and pcap_list_datalinks() if
called on a non-activated pcap_t.  Pre-pcap_create() code will never see
PCAP_ERROR_NOT_ACTIVATED, and code using pcap_create() will either not
see it (if it's correct) or is already getting bogus results (if it's
not correct).

Document that while we're at it.