Discourage the use of a zero timeout.
On BSD, OS X, and possibly Solaris 11 and later, it could cause your
application to wait indefinitely long for packets, which you probably
don't want.

On Linux, it could do any of a number of things, none of which will do
very much buffering of packets; if that's what you want, use a very
small timeout value (or use immediate mode).  Making it block forever is
difficult if not impossible.

On Solaris 10 or before, it's not clear from the bufmod(7m)
documentation what it would do; it could be made to wait indefinitely
long, or return packets immediately, or return after an
arbitrarily-chosen timeout, if we wanted to make it do so.

I'm too lazy to check what it'd do on various other OSes (other than the
ones that have no notion of a read timeout).