Don't treat all open errors as reasons not to list an interface.
Only treat "the capture mechanism doesn't know about this device" and
"you can't capture on the interface because it isn't up" as reasons not
to list it, as there's no point in showing those interfaces to users.
For all other reasons, show it to the user, so they can try to capture
on it and then find out why it doesn't work (e.g., you're not running
with sufficient permissions).