Fix some socket file descriptor leaks.
Thanks to Coverity for finding the first one.  The second one wasn't as
obviously a leak, so Coverity didn't find it; don't save the socket FD
in the pcap_t structure until we've succeeded in activating the pcap_t,
so that leaks such as that will be more obvious to Coverity.