Handle a timeout value of 0 better with TPACKET_V3.
If the user specified a timeout value of 0, and we don't have any
packets to return from pcap_read_linux_mmap_v3(), go back and wait for
packets.  Yes, this means waiting indefinitely, but that's what the
documentation used to say, and what we're making it say again; if this
is a problem, *don't use a timeout value of 0*.

This should fix GitHub libpcap issue #350.