Ask that people inform us of additions to BPF.
At minimum, we'd like to keep track of new instructions added, so that
others don't use the same encoding for different operations.  Ideally,
we'd like to update our implementation to support the new instructions
and encourage other implementations to add them as well.

Also remove the "from Linux" note on BPF_MOD and BPF_XOR, as the Linux
kernel BPF implementation isn't the only one that has them now, given
that we have them.