Fix the swapping of isochronous descriptors.
From looking at the Linux USB monitoring code:

With the old header, there are no isochronous descriptors after the
header.

With the new header, the actual number of descriptors in the header is
not s.iso.numdesc, it's ndesc - only the first N descriptors, for some
value of N, are put into the header, and ndesc is set to the actual
number copied.  In addition, if s.iso.numdesc is negative, no
descriptors are captured, and ndesc is set to 0.