Eliminate some unneeded variables and move some comments.
Eliminate orig_linktype; we were setting it but never using it.

Eliminnte orig_nl; we were setting it, but in the only places where we
were using it, it had the same value as off_nl, so just use off_nl.

Move some comments next to the code to which they're actually referring
(don't speak of changing offsets before we actually change the offsets).