Include fixed offset when computing variable MAC headers.
Some filters (such as VLAN) increment off_macpl as they are parsed.
In the fixed MAC header case, this is handled properly since it is
simply incorporated into the header length. However, currently with
variable length headers we only look at the variable component and ignore
the fixed part, so the VLAN stack can't be handled.

This includes both pieces for variable headers; in cases where there
are no VLAN filters, off_macpl is zero so things still work.