Update constants and variable name to reflect reality.
The offset of the MAC-layer payload is, in the general case, now the
sum of a variable value, computed based on the contents of various parts
of the packet, and a constant value.  Rename off_macpl to
off_macpl_constant_part to indicate that it's the constant value in
question, not necessarily the full offset, and update a lot of comments
to reflect that reality.