Rename some variables and enum values to make things a bit clearer.
Replace "ll"/"LL" with "linkhdr"/"LINKHDR", matching "linkpl"/"LINKPL",
as it refers to the offset to the link-layer header.

Use "prev", not "outer", for the link-layer information for the layer in
which the current one is encapsulated (if there's any encapsulation);
that name works better if there are multiple layers of encapsulation.