Rename OR_NET and OR_NET_NOSNAP to OR_LINKPL and OR_LINKPL_NOSNAP.
It really means "payload of the link layer", where "link layer" includes
any 802.2 LLC headers, and is used for fields not at the "network
layer".  The old OR_LINKPL was no longer used - either OR_LLC is used,
for testing fields in the 802.2 LLC header, or OR_MPLSPL is used.