Include constant_part when loading IP header length.
When calculating the IP header length, we only use the variable
part of the link header to locate the IP header but we also need
to use any constant components.

Note that the constant part is not included in the total returned -
it is only used to actually locate the IP header. This is consistent
with other behavior.