Mark radio link headers as variable.
Previously, not all variable length offsets had an is_variable flag
- this knowledge was essentially encoded in the generating logic.
Now that the offset code has been generalized, the flag has been
included for these cases as well and it needs to be set appropriately.
The radio code was missing it for the link headers so this adds it.