Initialize off_outermostlinkhdr and off_prevlinkhdr at start.
Currently off_outermostlinkhdr is always set to zero rather than
the true offsets. It is currently only used in one place where
the correct offset is zero but this is not semantically correct.
It is instead better to initialize the offsets to the initial
values computed for the link type.