Compute variable link header if we are computing the link payload.
The link payload is an absolute offset from the start of the packet,
meaning that it should include any variable link prefix as well.
However, currently if somebody asks for the payload but not the link
header, the latter is not computed and is skipped. This adds a
dependency between the two, so that we always calculate the full value.