Fix documentation.
Sadly, there's no fopen() mode that says "create this if it doesn't
exist, *leave it alone if it does*, open the file for reading and
writing, and *don't* force all writes to append to the file"; for now,
we just go with the semantics of "r+", which means it fails if it
doesn't exist.  That's what Mark Johnston's patch, from which the
earlier change was derived, did.