bpf: if generating code for vlan keyword also check for non-offloaded case
Our bpf compiler should also handle cases if vlan rx-offload is turned off on
the device or it did not happen because 802.1q encapsulated frame contained
non-standard vlan tpid value. Under described circumstances BPF code which
checks if tag is present and has desired value fails, thus packet is dropped.

This commit alters behavior of compiler such that fallback case is generated. If
first check fails then we check for vlan tpid and vlan tci in the packet it
self. If they are present we then store displacement to memory cell 0, which
needs to be applied to all offsets generated from later vlan checks.