Don't timeout poll() with TPACKET_V3 in 3.19 and later.
We don't need to do the workaround for not getting wakeups when buffers
are handed to us, nor do we have to do a timeout in poll() for the
benefit of programs that expect the timeout to expire even if no packets
have arrived, so, with TPACKET_V3 in 3.19 and later, we always use a
timeout of -1 for poll().

This eliminates some spurious wakeups from poll(), due to the poll()
call *itself* timing out, that turn into spurious timeouts; see GitHub
issue #383.