Static variables considered harmful.
There's no reason *not* to allow two different threads in a process to
read from separate pcap_t's; the static variable used for handling pcap
file records with a caplen > the file's snapshot length wouldn't work if
that's done.

Instead, just read the first snaplen bytes into the buffer and then
read-and-discard the rest of the bytes of the record (we can't seek, as
we might be reading from a pipe).