Squelch a Clang static analyzer complaint, and fix a real-but-unlikely issue.
Avoid a not-a-real-problem warning from the Clang static analyzer.

Fix an real-but-unlikely (requires a file with a huge number of IDBs)
issue that I *thought* was the source of the static analyzer's complaint
but that, in fact, it apparently hadn't noticed, namely the possibility
of a multiplication overflowing.

Fix a typo while we're at it.