Fix up the statistics handling in WinPcap.
Have a pcap_stat structure in the private data structure, and use that
to count received packets, for the benefit of adapter types (such as DAG
cards on Windows) that don't provide that count.

Don't play structure pointer casting games; that's fragile, and might
get in the way of using PacketGetStatsEx() in the future.

(We really need to redo the statistics handling so that it supplies a
structure similar to a pcapng Interface Statistics block; that'd let us

	1) indicate which statistics are actually being supplied, rather
	   than supplying fake values of 0 for statistics not being
	   supplied;

	2) supply additional types of statistics (and more clearly
	   indicate what type of received-packet-count is being
	   supplied);

	3) handle statistics that have counters bigger than 32 bits;

and then implement pcap_stats() atop that underlying API.)