Don't write past the end of an old pcap_stat structure in pcap_stats().
We might be called from a program compiled with an older version of
WinPcap in which the structure doesn't have a ps_capt member, in which
case filling that member in will overwrite whatever is after that
structure, so don't do that.

Don't worry about the packet count we maintain - pcap_stats() ignored
it, just returning whatever PacketGetStats() returned, and
pcap_stats_ex() overwrote it with whatever PacketGetStatsEx().

Fix a parenthesis error while we're at it.