Pick up the "Win32 extensions", and add some new ones.
Pick up the "Win32 extension" routines, and declare them in <pcap/pcap.h>
if we're building for Win32, so Win32-Extensions.[ch] are no longer
necessary.

Make pcap_win32strerror() take a Win32 error code and a buffer pointer
as arguments, and fill in the buffer, assumed to be of size
PCAP_ERRBUF_SIZE+1.  That way, we're not filling in a static buffer, so
error reporting isn't a problem in multi-threaded code.  (No guarantees
that you can act on the *same* pcap_t in different threads, but, in most
cases, you should be able to act on *different* pcap_t's in different
threads.)

Add pcap_oid_get_request() and pcap_oid_set_request() to do OID
accesses; this removes the primary reason why callers would need to get
the ADAPTER structure for a pcap_t, so we just get rid of
pcap_get_adapter().