Add the remaining WinPcap-only ops to the ops vector.
That way, all pcap_t types can handle them in Win32, even if they just
reject the op because they don't support it.

Fill those ops in as appropriate.

Get rid of the now-unused "get adapter" op.

Nobody needs to use pcap_not_initialized() outside pcap.c -
pcap_create() initializes most of the ops to point to it - so make it
static.

Have pcap_get_airpcap_handle() supply an error string of "that's not an
AirPcap device" if the op returns NULL.