Add TurboCap support from WinPcap.
1) It's also apparently supported on Linux.

2) This cleans it up to match the way we're currently dealing with pcap
   modules.

3) This cleans it up to the way we're handling the Win32 extensions.

Fix the configure script so that, if a pcap module that is supported
only if the relevant support software is available is being configured
in, we *add* its source to the list of source files to compile, rather
than making it *the* source file to compile, in case there's more than
one of them.

Fix a typo in a comment.

Add a comment about TurboCap using DLT_PPI for Ethernet.

Move the struct pcap_stat used for pcap_stats_ex() into the pcap
structure on Windows, as it's used by multiple modules.

Remove the stuff that handles TurboCap from pcap-win32.c, as the
routines that did that are now just module ops, so we call the TurboCap
versions for TurboCap devices.

Remove the stuff that handles remote captures from there as well; that
will eventually be handles by modules (plural) for the remote capture
protocols (plural).

Make all of the fields added to struct pcap_stat dependent on _WIN32 and
HAVE_REMOTE being defined.  We don't want any of them for UN*X, as that
would potentially break binary compatibility; we make them all dependent
on HAVE_REMOTE being defined because that's what WinPcap currently does
(yes, that's a bad idea, but we'll be fixing this by deprecating that
structure as well as pcap_stats() and pcap_stats_ex() in favor of
something a bit more pcapng-like, which would be more extensible).