Remove platform checks for various sniffing hardware APIs.
Those checks made sense when particular system pcap modules did checks
for particular third-party devices, so if your system's appropriate pcap
module didn't check for it, there's no point in asking for it.

Now, those checks come from a table in pcap.c, so they could be done on
any system.  Either we find the API headers and libraries or we don't.
If we do, we might as well assume there's an implementation for them on
the platform for which we're building.  If we don't, the configure
script will fail because of that, regardless of the platform.