Declare pcap_wsockinit() in pcap-int.h.
We probably shouldn't be exporting something called wsockinit() at all,
as it's too likely to collide with routines in code that calls us.  I
don't know whether Wine implemented it in order to run Windows programs
that call it, or if there actually *are* such programs, but we'll leave
it around for now.

Call pcap_wsockinit() in pcap_compile(), and don't bother declaring it;
rely on it being declared in pcap-int.h, so we run less of a risk of the
declaration and definition getting out of sync (not that we're likely to
change its signature, but...).

In the definitions of it and wsockinit(), explicitly note that they take
no arguments.