Just use a PACKET structure on the stack in the read routines.
There's no state that needs to be preserved between calls, so just use
an on-stack structure and fill it in before each PacketReceivePacket()
call.

That's one less WinPcap-specific data structure in the pcap_t structure
(and the PACKET structure itself wasn't in the pcap_t structure, so it's
not as if you got any cache locality improvement from it).