Have better APIs for the OID get/set request calls.
Have them take the OID, the buffer, and the length as arguments, and,
for pcap-win32.c, have the routines for those calls allocate a
PACKET_OID_DATA, fill it in as appropriate, and copy the data back.

Yeah, it's more work and another memory allocation, but it's still a
cleaner API and avoids some compiler warnings.