Add the pcap_file_type() function and the file_type member in pcap_t
This function is used to get the pcap file type when reading a pcap file.
The types are currently pcap and pcap-ng (PCAP_FILE_TYPE and
PCAP_NG_FILE_TYPE).

The first use of this function is to allow tcpdump to know the file type
and to do relevant version checks based on these types.

For that purpose, we need to move PCAP_NG_VERSION_MAJOR from sf-pcap-ng.c
to pcap/pcap.h.