Add a "getopt.h" header and include it instead of <unistd.h> on Windows.
We include <unistd.h> for getopt(), but, on Windows, we supply our own
getopt() (as it's a UN*X API), so supply our own header and include that
instead.

While we're at it, mark the copyright strings as unused (to squelch
compiler warnings), and don't include the headers for select() or poll()
in capturetest.c, as we don't use those APIs there (that's what
selpolltest(), which is UN*X-only, is for).  <unistd.h> or "getopt.h"
should declare the getopt() external variables, so don't declare them
ourselves.