Attempt to get UN*X compilers to work like MSVC.
If the compiler has a flag to specify that symbols should be hidden
outside the shared library unless explicitly marked as visible, specify
it.  Make PCAP_API mark a symbol as visible if the compiler supports it.

If it doesn't have that, it probably also lacks the ability to mark
symbols as hidden, so don't bother with explicitly marking symbols as
hidden.