Go back to using PCAP_API on Windows.
Using an exports file for all platforms turns out to be messier than
telling the compiler about exports, and telling the compiler about
exports gives the compiler some hints allowing it to generate better
code.  It does mean we don't get to limit exports on platforms where
that can only be done at link time, at least not without having some
script to extract the names of the relevant symbols from the appropriate
header files or maintaining a separate symbol file, but that leaves only
HP-UX, AIX, and perhaps some older versions of the other platforms
unable to limit exports.