Have routines to set various internal debugging flags.
Those flags are available only in versions of libpcap built with support
for them; they're not intended to be real APIs, they're intended to be
used if you're doing debugging of libpcap itself, and it's expected that
you'll be linking with a program that knows about them and that declares
them itself.

There's no clean way of discouraging their use except in special cases -
it's really ugly providing them in a header if they're not going to be
present by default, so this is about the best we can do.

This is, at least, cleaner than exporting them as data.

Clean up some spacing while we're at it.