Split out the UN*X flags to pcap flags mapping.
Add an if_flags_to_pcap_flags() routine to map the interface name (if
necessary) and IFF_ flags to PCAP_IF_ flags, and call that in the
platform-specific routines.  That means that the calls to add to the
interface list don't need to use IFF_ flags, which are UN*Xisms.

Move the routines that update the interface list to fad-helpers.c,
leaving just the legacy APIs in inet.c.

Update finalldevstest to print out all the flags.