Clean up hardware time stamp support.
Don't say that the "any" device supports any time stamp types; we can't
arrange that every device, including devices that show up in the future,
do hardware time stamping.

If we can ask about hardware time stamp support, don't say we support
hardware time stamping if the answer is "not for all packets".

If we get ERANGE from an attempt to set hardware time stamping, it means
"yes, we support hardware time stamping, but not for all packets
matching that filter"; the filter in question is "all packets", so here
it would mean "we can't do hardware time stamping for all incoming
packets", which means "we can't do time stamping for captured packets".