Change the way NFLOG version is determined
The NOT operation is inside the brackets because it only applies to the variable nfhdr->nflog_version, not to the expression.
This check is pretty straightforward but silences a compiler warning. The operator != is not the best usage here, so I replaced it with a solution that made his proof in Nmap project (I am currently working for Nmap organisation).