Let the platform decide how to check capturable interfaces.
(Git's annoying policy of expecting a short one-line description of every
change means that the first line isn't very explanatory.)

Make pcap_findalldevs_interfaces() take as an argument a function that's
used to check whether an interface can be captured on or not, rather
than doing the check by trying to open the device for capturing.

This lets pcap_findalldevs() find interfaces even if you don't have
permission to capture on them; that way, instead of users saying "why
isn't {tcpdump -D, tshark -D, dumpcap -D, Wireshark, etc.} showing me
any interfaces?", they'll say "why am I getting a 'you don't have
permissions' error when I try to capture on this interface?", which is a
better description of the underlying problem.

On some platforms, it also avoids a bunch of extra work when getting a
list of interfaces.