Clean up {DAG, Septel, Myricom SNF}-only builds.
In the core code, treat them similarly to other builds.  In the
configure script, ensure that no other pcap-XXX.c file is built, so we
have only the entries for the capture mechanism in question in the
tables for pcap_findalldevs() and pcap_create(), so that those routines
are the only ones called.  If XXX_ONLY is defined in pcap-XXX.c, define
a stub pcap_platform_finddevs() that finds no regular interfaces and a
stub pcap_create_interface() that fails with a "we only support XXX"
error message.