Check for, and squelch, signed vs. unsigned comparison errors.
Use -Wsign-compare if the compiler supports it.

If we're comparing an unsigned value with a signed value, and we've
already determined that the signed value is >= 0, just cast it to an
unsigned type.

Use 0xffffffffU instead of -1 as the "unknown"/"unset"/etc. value for
unsigned variables and fields.

Assign the result of str2tok() to an int, not a u_int, as it can return
-1.

Declare some variables and fields unsigned if they don't need to be
signed.  Make sure some arguments passed into "set" functions are
non-negative (and otherwise not invalid).

In the BPF optimizer, cast the "constant" fields of a struct block to an
unsigned type - the actual constant field of a BPF instruction is
unsigned.

Also, cast away one warning from MSVC.