Re-impose some of Winsock's limitations on sock_recv().
In Winsock, recv() takes an int as the size argument and returns an int
as a return value, so it can't read more than INT_MAX bytes.  Have it
take a size_t as an argument (so you don't have to cast away warnings
about narrowing sizeof values), return an error if it's bigger than
INT_MAX (which it can be even on ILP32 platforms, as size_t is
unsigned), and have the count of remaining bytes be an int.