Give a better discussion of possible errors.
Give each warning and error code its own paragraph.

Note that, in the future, additional warning and error codes may be
added, so programs should treat *all* positive return codes as warnings
and *all* negative return codes as errors, and, if the warning or error
code isn't one the program specifically checks for, and isn't
PCAP_WARNING or PCAP_ERROR, it should use pcap_statustostr() to report
the warning or error.

(Yes, we've added additional return codes in the past, subsequent to
release 1.0, so it's not as if we're changing anything with this
announcement.)