Add a wrapper for strtok_r().
Have a pcap_strtok_r(), which behaves like the standard UNIX strtok_r().

If we have strtok_r(), pcap_strtok_r is just #defined to be strtok_r.

If we're building with MSVC, pcap_strtok_r is just #defined to be
strtok_s.

Otherwise, we build our own, taken from FreeBSD.

Use it in sockutils.c, to avoid warnings and to make sure it's
thread-safe (the Single UNIX Specification does not guarantee that
strtok() is thread-safe - it could, but it isn't required to, use
thread-local storage for the state).